// frontend/src/api/generated/models/pieceRatePayload.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PieceRatePayload {
  product_model_id: string;
  stage_name: string;
  rate_amount: number | string;
  effective_date: string;
  operator_id?: string | null;
  location?: string;
  operator_status?: string;
  change_reason: string;
}