// frontend/src/api/generated/models/pieceRateResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface PieceRateResponse {
  id: string;
  product_model_id: string;
  stage_name: string;
  /** @pattern ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$ */
  rate_amount: string;
  effective_date: string;
  operator_id?: string | null;
}