// frontend/src/api/generated/models/productModelResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Manufacturing Information System API
 * API untuk aplikasi core MIS.
 * OpenAPI spec version: 1.0.0
 */

export interface ProductModelResponse {
  id: string;
  code: string;
  name: string;
  description: string;
  is_active: boolean;
}